sched/core: always dequeue dl task when changing parameters
The optimization of updating task parameters without moving the task
improves the performance for rt tasks, but causes problems in the
bandwidth accounting for dl tasks.

More specifically, dl tasks need to be enqueued and dequeued to properly
update the bandwidth of the runqueue, so this workaround disables that
optimization for dl tasks.