fuse: Add support for passthrough read/write
Add support for filesystem passthrough read/write of files when enabled
in userspace through the option FUSE_PASSTHROUGH.

There are filesystems based on FUSE that perform checks, or enforce
special policies, or perform other kind of decision making when certain
operations are performed on their files. Android, for example, uses FUSE
to enforce some fine-grained access policies that also depend on the
contents of the files.
Sometimes it happens that at "open" or "create" time a file is
identified as not requiring additional read/write checkings, thus FUSE
would simply act as a passive bridge between the FUSE and the lower
filesystem.
Redirecting some of the read/write operations to the userspace FUSE
daemon still introduces some overhead that this patch aims at
minimizing.

The target file object towards which the read/write operations will be
forwarded in passthrough mode is handled in a way that reminds the file
descriptor passing with sockets:
- the FUSE daemon opens the target file in the lower filesystem;
- the file descriptor is passed to the kernel via /dev/fuse;
- the kernel gets the file pointer through the opened files of the
  current process and stores it in an additional file pointer field in
  fuse_file.
From now all the read/write operations performed by the process that
opened the FUSE file will be directly applied to the file pointer
pointing at the lower filesystem's file.

When the FUSE_PASSTHROUGH capability is enabled, the FUSE daemon may
decide while handling the "open" operation, if the given file can be
accessed by that process in "passthrough" mode, meaning that all the
further read and write operations would be forwarded by the kernel
directly to the lower filesystem rather than to the FUSE daemon. All
requests that aren't read or write are still handled by the userspace
code.

This allows for an improved performance on reads and writes, especially
in the case of reads at random offsets, for which no (readahead)
caching mechanism would help.

Change log:

v6:
Removed get_file() and fputs() during read/write, as the file ref count
is already handled at fuse_file creation/deletion.
[Identified by Jann Horn]

v5:
Fix the check when setting the passthrough file
[Found when testing by Mike Shal]

v3 and v4:
Use the fs_stack_depth to prevent further stacking and a minor fix
[Suggested by Jann Horn]

v2:
Changed the feature name to passthrough from stacked_io
[Proposed by Linus Torvalds]

Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>
Signed-off-by: Alessio Balsini <balsini@android.com>