fuse: Add support for passthrough read/write
Add support for filesystem passthrough read/write of files when enabled in
userspace through the option FUSE_PASSTHROUGH.

There are filesystems based on FUSE that are intended to enforce special
policies or trigger complicate decision makings at the file operations level.
Android, for example, uses FUSE to enforce fine-grained access policies that
also depend on the file contents.
Sometimes it happens that at open or create time a file is identified as not
requiring additional checks for consequent reads/writes, thus FUSE would simply
act as a passive bridge between the process accessing the FUSE filesystem and
the lower filesystem. Splicing and caching help reducing the FUSE overhead, but
there are still read/write operations forwarded to the userspace FUSE daemon
that could be avoided.

When the FUSE_PASSTHROUGH capability is enabled, the FUSE daemon may decide
while handling the open or create operations, if the given file can be accessed
in passthrough mode, meaning that all the further read and write operations
would be forwarded by the kernel directly to the lower filesystem rather than
to the FUSE daemon. All requests that are not reads or writes are still handled
by the userspace code.
This allows for improved performance on reads and writes, especially in the
case of reads at random offsets, for which no (readahead) caching mechanism
would help.
Benchmarks show improved performance that is close to native filesystem access
when doing massive manipulations on a single opened file, especially in the
case of random reads, for which the bandwidth increased by almost 2X or
sequential writes for which the improvement is close to 3X.

The creation of this direct connection (passthrough) between FUSE file objects
and file objects in the lower filesystem happens in a way that reminds of
passing file descriptors via sockets:
- a process opens a file handled by FUSE, so the kernel forwards the request to
  the FUSE daemon;
- the FUSE daemon opens the target file in the lower filesystem, getting its
  file descriptor;
- the file descriptor is passed to the kernel via /dev/fuse;
- the kernel gets the file pointer navigating through the opened files of the
  "current" process and stores it in an additional field in the fuse_file owned
  by the process accessing the FUSE filesystem.
From now all the read/write operations performed by that process will be
redirected to the file pointer pointing at the lower filesystem's file.
Handling asynchronous IO is done by creating separate AIO requests for the
lower filesystem that will be internally tracked by FUSE, that intercepts and
propagates their completion through an internal ki_completed callback similar
to the current implementation of overlayfs.

Original-patch-by: Nikhilesh Reddy <reddyn@codeaurora.org>
Signed-off-by: Alessio Balsini <balsini@android.com>