fuse: Definitions and ioctl() for passthrough
Introduce the new FUSE passthrough ioctl(), which allows userspace to
specify a direct connection between a FUSE file and a lower file system
file.
Such ioctl() requires userspace to specify the file descriptor of one of
its opened files through a fuse_passthrough_out data structure, which
includes extra fields available for future extensions.
The file pointer obtained via the file descriptor passed to the ioctl()
will be copied to the fuse_file data structure, so that the link between
FUSE and lower file system is consolidated.

In order for the passthrough mode to be successfully activated, the lower
file system file must implement both read_ and write_iter file operations.
This extra check avoids special pseudofiles to be targets for this feature.
An additional enforced limitation is that when FUSE passthrough is enabled,
no further file system stacking is allowed.

Signed-off-by: Alessio Balsini <balsini@android.com>