fuse: Introduce synchronous read and write for passthrough
All the read and write operations performed on fuse_files which have the
passthrough feature enabled are forwarded to the associated lower file
system file.

Sending the request directly to the lower file system avoids the userspace
round-trip that, because of possible context switches and additional
operations might reduce the overall performance, especially in those cases
where caching doesn't help, for example in reads at random offsets.

The connection between the fuse_files and lower file system file is created
through the FUSE_DEV_IOC_PASSTHROUGH_OPEN ioctl() and can be verified if
the passthrough_filp pointer in the fuse_file is not null.
After the read/write operation is completed, both the access/modify
operations must be notified (and propagated) to the lower file system,
while this is not necessary for the FUSE file system because automatically
triggered by the read/write caller.
This change only implements synchronous requests in passthrough, returning
an error in the case of ansynchronous operations, yet covering the majority
of the use cases.

Signed-off-by: Alessio Balsini <balsini@android.com>