fuse: Handle AIO read and write in passthrough
Extend the passthrough feature by handling asynchronous IO both for read
and write operations.
When an incoming AIO request targeting a FUSE file with passthrough
functionality enabled, a new identical AIO request is created, the file
pointer of which is updated with the file pointer of the lower file system,
and the complition handler is updated with a special AIO passthrough
handler.
The lower file system AIO request is allocated in dynamic kernel memory
and, when it completes, the allocated memory is freed and the completion
signal is propagated to the FUSE AIO request by triggering its completion
callback as well.

Signed-off-by: Alessio Balsini <balsini@android.com>